Add generic attributes helper
Replaced processClass, processRequired, processId and processPlaceholder
functionality with a generic processAttributes function.

Some known HTML5 attributes will be treated as boolean, the rest will be
added in `attribute="value"` form. This also makes it possible to add
arbitrary attributes to form fields.

There is one backwards incompatible change in here: if you specify a
class, the default classes (like `form-control`) are overridden. So if
you depend on these, you have to add them to the field configuration.
Example:

```
{{text_field 'title' required=true class="form-control custom"}}
```